Feature/add rules protection

+++ backend/src/resolvers/Mutation/System/webTool.ts
@@ -1,19 +1,29 @@
 import db from '../../../config/db';
 import { webTool as getWebTool } from '../../Query/System/webTool';
 
-const createWebTool = async (_, { data }) => {
+const createWebTool = async (_, { data }, ctx) => {
+    const webClient = ctx.getWebClient();
+    if (!webClient || !ctx.hasPermission('RULE_SUPERADMIN')) {
+        return new Error('access denied');
+    }
+
     try {
         return db('web_tool').insert(data)
-            .then(id => getWebTool(_, { filter: { id } }))
+            .then(id => getWebTool(_, { filter: { id } }, ctx))
             .catch(e => new Error(e.sqlMessage));
     } catch (e) {
         throw new Error(e.sqlMessage);
     }
 }
 
-const updateWebTool = async (_, { filter, data }) => {
+const updateWebTool = async (_, { filter, data }, ctx) => {
+    const webClient = ctx.getWebClient();
+    if (!webClient || !ctx.hasPermission('RULE_SUPERADMIN')) {
+        return new Error('access denied');
+    }
+
     try {
-        const webTool = await getWebTool(_, { filter });
+        const webTool = await getWebTool(_, { filter }, ctx);
         if (webTool) {
             const { id } = webTool;
             await db('web_tool').where({ id }).update(data)
@@ -24,9 +34,14 @@ const updateWebTool = async (_, { filter, data }) => {
     }
 }
 
-const deleteWebTool = async (_, { filter }) => {
+const deleteWebTool = async (_, { filter }, ctx) => {
+    const webClient = ctx.getWebClient();
+    if (!webClient || !ctx.hasPermission('RULE_SUPERADMIN')) {
+        return new Error('access denied');
+    }
+
     try {
-        const webTool = await getWebTool(_, { filter });
+        const webTool = await getWebTool(_, { filter }, ctx);
         if (webTool) {
             const { id } = webTool;
             await db('web_tool').where({ id }).delete();
